requirements.txt
PyJWT
requests
flask
dotenv